+++ sdk/netapp/azure-mgmt-netapp/azure/mgmt/netapp/models/_models.py
@@ -10,6 +10,24 @@
 # --------------------------------------------------------------------------
 
 from msrest.serialization import Model
+from msrest.exceptions import HttpOperationError
+
+
+class AccountEncryption(Model):
+    """Encryption settings.
+
+    :param key_source: Encryption Key Source. Possible values are:
+     'Microsoft.NetApp'. Possible values include: 'Microsoft.NetApp'
+    :type key_source: str or ~azure.mgmt.netapp.models.KeySource
+    """
+
+    _attribute_map = {
+        'key_source': {'key': 'keySource', 'type': 'str'},
+    }
+
+    def __init__(self, **kwargs):
+        super(AccountEncryption, self).__init__(**kwargs)
+        self.key_source = kwargs.get('key_source', None)
 
 
 class ActiveDirectory(Model):
@@ -40,7 +58,7 @@ class ActiveDirectory(Model):
      registered as a computer account in the AD and used to mount volumes
     :type smb_server_name: str
     :param organizational_unit: The Organizational Unit (OU) within the
-     Windows Active Directory
+     Windows Active Directory. Default value: "CN=Computers" .
     :type organizational_unit: str
     :param site: The Active Directory site the service will limit Domain
      Controller discovery to
@@ -70,6 +88,9 @@ class ActiveDirectory(Model):
      given SeSecurityPrivilege privilege (Needed for SMB Continuously available
      shares for SQL). A list of unique usernames without domain specifier
     :type security_operators: list[str]
+    :param ldap_over_tls: Specifies whether or not the LDAP traffic needs to
+     be secured via TLS.
+    :type ldap_over_tls: bool
     """
 
     _validation = {
@@ -99,6 +120,7 @@ class ActiveDirectory(Model):
         'aes_encryption': {'key': 'aesEncryption', 'type': 'bool'},
         'ldap_signing': {'key': 'ldapSigning', 'type': 'bool'},
         'security_operators': {'key': 'securityOperators', 'type': '[str]'},
+        'ldap_over_tls': {'key': 'ldapOverTLS', 'type': 'bool'},
     }
 
     def __init__(self, **kwargs):
@@ -111,7 +133,7 @@ class ActiveDirectory(Model):
         self.status = None
         self.status_details = None
         self.smb_server_name = kwargs.get('smb_server_name', None)
-        self.organizational_unit = kwargs.get('organizational_unit', None)
+        self.organizational_unit = kwargs.get('organizational_unit', "CN=Computers")
         self.site = kwargs.get('site', None)
         self.backup_operators = kwargs.get('backup_operators', None)
         self.kdc_ip = kwargs.get('kdc_ip', None)
@@ -120,6 +142,7 @@ class ActiveDirectory(Model):
         self.aes_encryption = kwargs.get('aes_encryption', None)
         self.ldap_signing = kwargs.get('ldap_signing', None)
         self.security_operators = kwargs.get('security_operators', None)
+        self.ldap_over_tls = kwargs.get('ldap_over_tls', None)
 
 
 class AuthorizeRequest(Model):
@@ -166,6 +189,8 @@ class Backup(Model):
     :type label: str
     :ivar backup_type: Type of backup adhoc or scheduled
     :vartype backup_type: str
+    :ivar failure_reason: Failure reason
+    :vartype failure_reason: str
     """
 
     _validation = {
@@ -178,6 +203,7 @@ class Backup(Model):
         'provisioning_state': {'readonly': True},
         'size': {'readonly': True},
         'backup_type': {'readonly': True},
+        'failure_reason': {'readonly': True},
     }
 
     _attribute_map = {
@@ -191,6 +217,7 @@ class Backup(Model):
         'size': {'key': 'properties.size', 'type': 'long'},
         'label': {'key': 'properties.label', 'type': 'str'},
         'backup_type': {'key': 'properties.backupType', 'type': 'str'},
+        'failure_reason': {'key': 'properties.failureReason', 'type': 'str'},
     }
 
     def __init__(self, **kwargs):
@@ -205,6 +232,7 @@ class Backup(Model):
         self.size = None
         self.label = kwargs.get('label', None)
         self.backup_type = None
+        self.failure_reason = None
 
 
 class BackupPatch(Model):
@@ -227,6 +255,8 @@ class BackupPatch(Model):
     :type label: str
     :ivar backup_type: Type of backup adhoc or scheduled
     :vartype backup_type: str
+    :ivar failure_reason: Failure reason
+    :vartype failure_reason: str
     """
 
     _validation = {
@@ -235,6 +265,7 @@ class BackupPatch(Model):
         'provisioning_state': {'readonly': True},
         'size': {'readonly': True},
         'backup_type': {'readonly': True},
+        'failure_reason': {'readonly': True},
     }
 
     _attribute_map = {
@@ -245,6 +276,7 @@ class BackupPatch(Model):
         'size': {'key': 'properties.size', 'type': 'long'},
         'label': {'key': 'properties.label', 'type': 'str'},
         'backup_type': {'key': 'properties.backupType', 'type': 'str'},
+        'failure_reason': {'key': 'properties.failureReason', 'type': 'str'},
     }
 
     def __init__(self, **kwargs):
@@ -256,6 +288,7 @@ class BackupPatch(Model):
         self.size = None
         self.label = kwargs.get('label', None)
         self.backup_type = None
+        self.failure_reason = None
 
 
 class BackupPolicy(Model):
@@ -701,11 +734,53 @@ class CheckAvailabilityResponse(Model):
 
 
 class CloudError(Model):
-    """CloudError.
+    """An error response from the service.
+
+    :param error: Cloud error body.
+    :type error: ~azure.mgmt.netapp.models.CloudErrorBody
     """
 
     _attribute_map = {
+        'error': {'key': 'error', 'type': 'CloudErrorBody'},
     }
+
+    def __init__(self, **kwargs):
+        super(CloudError, self).__init__(**kwargs)
+        self.error = kwargs.get('error', None)
+
+
+class CloudErrorException(HttpOperationError):
+    """Server responsed with exception of type: 'CloudError'.
+
+    :param deserialize: A deserializer
+    :param response: Server response to be deserialized.
+    """
+
+    def __init__(self, deserialize, response, *args):
+
+        super(CloudErrorException, self).__init__(deserialize, response, 'CloudError', *args)
+
+
+class CloudErrorBody(Model):
+    """An error response from the service.
+
+    :param code: An identifier for the error. Codes are invariant and are
+     intended to be consumed programmatically.
+    :type code: str
+    :param message: A message describing the error, intended to be suitable
+     for display in a user interface.
+    :type message: str
+    """
+
+    _attribute_map = {
+        'code': {'key': 'code', 'type': 'str'},
+        'message': {'key': 'message', 'type': 'str'},
+    }
+
+    def __init__(self, **kwargs):
+        super(CloudErrorBody, self).__init__(**kwargs)
+        self.code = kwargs.get('code', None)
+        self.message = kwargs.get('message', None)
 
 
 class DailySchedule(Model):
@@ -1075,6 +1150,10 @@ class NetAppAccount(Model):
     :vartype provisioning_state: str
     :param active_directories: Active Directories
     :type active_directories: list[~azure.mgmt.netapp.models.ActiveDirectory]
+    :param encryption: Encryption settings
+    :type encryption: ~azure.mgmt.netapp.models.AccountEncryption
+    :ivar system_data: The system meta data relating to this resource.
+    :vartype system_data: ~azure.mgmt.netapp.models.SystemData
     """
 
     _validation = {
@@ -1083,6 +1162,7 @@ class NetAppAccount(Model):
         'name': {'readonly': True},
         'type': {'readonly': True},
         'provisioning_state': {'readonly': True},
+        'system_data': {'readonly': True},
     }
 
     _attribute_map = {
@@ -1093,6 +1173,8 @@ class NetAppAccount(Model):
         'tags': {'key': 'tags', 'type': '{str}'},
         'provisioning_state': {'key': 'properties.provisioningState', 'type': 'str'},
         'active_directories': {'key': 'properties.activeDirectories', 'type': '[ActiveDirectory]'},
+        'encryption': {'key': 'properties.encryption', 'type': 'AccountEncryption'},
+        'system_data': {'key': 'systemData', 'type': 'SystemData'},
     }
 
     def __init__(self, **kwargs):
@@ -1104,6 +1186,8 @@ class NetAppAccount(Model):
         self.tags = kwargs.get('tags', None)
         self.provisioning_state = None
         self.active_directories = kwargs.get('active_directories', None)
+        self.encryption = kwargs.get('encryption', None)
+        self.system_data = None
 
 
 class NetAppAccountPatch(Model):
@@ -1126,6 +1210,8 @@ class NetAppAccountPatch(Model):
     :vartype provisioning_state: str
     :param active_directories: Active Directories
     :type active_directories: list[~azure.mgmt.netapp.models.ActiveDirectory]
+    :param encryption: Encryption settings
+    :type encryption: ~azure.mgmt.netapp.models.AccountEncryption
     """
 
     _validation = {
@@ -1143,6 +1229,7 @@ class NetAppAccountPatch(Model):
         'tags': {'key': 'tags', 'type': '{str}'},
         'provisioning_state': {'key': 'properties.provisioningState', 'type': 'str'},
         'active_directories': {'key': 'properties.activeDirectories', 'type': '[ActiveDirectory]'},
+        'encryption': {'key': 'properties.encryption', 'type': 'AccountEncryption'},
     }
 
     def __init__(self, **kwargs):
@@ -1154,6 +1241,7 @@ class NetAppAccountPatch(Model):
         self.tags = kwargs.get('tags', None)
         self.provisioning_state = None
         self.active_directories = kwargs.get('active_directories', None)
+        self.encryption = kwargs.get('encryption', None)
 
 
 class Operation(Model):
@@ -1694,6 +1782,46 @@ class SnapshotPolicyVolumeList(Model):
         self.value = kwargs.get('value', None)
 
 
+class SystemData(Model):
+    """Metadata pertaining to creation and last modification of the resource.
+
+    :param created_by: The identity that created the resource.
+    :type created_by: str
+    :param created_by_type: The type of identity that created the resource.
+     Possible values include: 'User', 'Application', 'ManagedIdentity', 'Key'
+    :type created_by_type: str or ~azure.mgmt.netapp.models.CreatedByType
+    :param created_at: The timestamp of resource creation (UTC).
+    :type created_at: datetime
+    :param last_modified_by: The identity that last modified the resource.
+    :type last_modified_by: str
+    :param last_modified_by_type: The type of identity that last modified the
+     resource. Possible values include: 'User', 'Application',
+     'ManagedIdentity', 'Key'
+    :type last_modified_by_type: str or
+     ~azure.mgmt.netapp.models.CreatedByType
+    :param last_modified_at: The timestamp of resource last modification (UTC)
+    :type last_modified_at: datetime
+    """
+
+    _attribute_map = {
+        'created_by': {'key': 'createdBy', 'type': 'str'},
+        'created_by_type': {'key': 'createdByType', 'type': 'str'},
+        'created_at': {'key': 'createdAt', 'type': 'iso-8601'},
+        'last_modified_by': {'key': 'lastModifiedBy', 'type': 'str'},
+        'last_modified_by_type': {'key': 'lastModifiedByType', 'type': 'str'},
+        'last_modified_at': {'key': 'lastModifiedAt', 'type': 'iso-8601'},
+    }
+
+    def __init__(self, **kwargs):
+        super(SystemData, self).__init__(**kwargs)
+        self.created_by = kwargs.get('created_by', None)
+        self.created_by_type = kwargs.get('created_by_type', None)
+        self.created_at = kwargs.get('created_at', None)
+        self.last_modified_by = kwargs.get('last_modified_by', None)
+        self.last_modified_by_type = kwargs.get('last_modified_by_type', None)
+        self.last_modified_at = kwargs.get('last_modified_at', None)
+
+
 class Vault(Model):
     """Vault information.
 
@@ -1773,7 +1901,8 @@ class Volume(Model):
     :param export_policy: exportPolicy. Set of export policy rules
     :type export_policy:
      ~azure.mgmt.netapp.models.VolumePropertiesExportPolicy
-    :param protocol_types: protocolTypes. Set of protocol types
+    :param protocol_types: protocolTypes. Set of protocol types, default
+     NFSv3, CIFS fro SMB protocol
     :type protocol_types: list[str]
     :ivar provisioning_state: Azure lifecycle management
     :vartype provisioning_state: str
@@ -1789,8 +1918,9 @@ class Volume(Model):
     :param subnet_id: Required. The Azure Resource URI for a delegated subnet.
      Must have the delegation Microsoft.NetApp/volumes
     :type subnet_id: str
-    :param mount_targets: mountTargets. List of mount targets
-    :type mount_targets: list[~azure.mgmt.netapp.models.MountTargetProperties]
+    :ivar mount_targets: mountTargets. List of mount targets
+    :vartype mount_targets:
+     list[~azure.mgmt.netapp.models.MountTargetProperties]
     :param volume_type: What type of volume is this
     :type volume_type: str
     :param data_protection: DataProtection. DataProtection type volumes
@@ -1800,14 +1930,15 @@ class Volume(Model):
     :param is_restoring: Restoring
     :type is_restoring: bool
     :param snapshot_directory_visible: If enabled (true) the volume will
-     contain a read-only .snapshot directory which provides access to each of
-     the volume's snapshots (default to true).
+     contain a read-only snapshot directory which provides access to each of
+     the volume's snapshots (default to true). Default value: True .
     :type snapshot_directory_visible: bool
     :param kerberos_enabled: Describe if a volume is KerberosEnabled. To be
      use with swagger version 2020-05-01 or later. Default value: False .
     :type kerberos_enabled: bool
-    :param security_style: The security style of volume. Possible values
-     include: 'ntfs', 'unix'
+    :param security_style: The security style of volume, default unix,
+     defaults to ntfs for dual protocol or CIFS protocol. Possible values
+     include: 'ntfs', 'unix'. Default value: "unix" .
     :type security_style: str or ~azure.mgmt.netapp.models.SecurityStyle
     :param smb_encryption: Enables encryption for in-flight smb3 data. Only
      applicable for SMB/DualProtocol volume. To be used with swagger version
@@ -1818,8 +1949,11 @@ class Volume(Model):
      False .
     :type smb_continuously_available: bool
     :param throughput_mibps: Maximum throughput in Mibps that can be achieved
-     by this volume.
+     by this volume.  Default value: 0 .
     :type throughput_mibps: float
+    :param encryption_key_source: Encryption Key Source. Possible values are:
+     'Microsoft.NetApp'
+    :type encryption_key_source: str
     """
 
     _validation = {
@@ -1835,7 +1969,8 @@ class Volume(Model):
         'backup_id': {'max_length': 36, 'min_length': 36, 'pattern': r'^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}|(\\?([^\/]*[\/])*)([^\/]+)$'},
         'baremetal_tenant_id': {'readonly': True},
         'subnet_id': {'required': True},
-        'throughput_mibps': {'maximum': 4500, 'minimum': 1, 'multiple': 0.001},
+        'mount_targets': {'readonly': True},
+        'throughput_mibps': {'maximum': 4500, 'minimum': 0, 'multiple': 0.001},
     }
 
     _attribute_map = {
@@ -1865,6 +2000,7 @@ class Volume(Model):
         'smb_encryption': {'key': 'properties.smbEncryption', 'type': 'bool'},
         'smb_continuously_available': {'key': 'properties.smbContinuouslyAvailable', 'type': 'bool'},
         'throughput_mibps': {'key': 'properties.throughputMibps', 'type': 'float'},
+        'encryption_key_source': {'key': 'properties.encryptionKeySource', 'type': 'str'},
     }
 
     def __init__(self, **kwargs):
@@ -1885,16 +2021,17 @@ class Volume(Model):
         self.backup_id = kwargs.get('backup_id', None)
         self.baremetal_tenant_id = None
         self.subnet_id = kwargs.get('subnet_id', None)
-        self.mount_targets = kwargs.get('mount_targets', None)
+        self.mount_targets = None
         self.volume_type = kwargs.get('volume_type', None)
         self.data_protection = kwargs.get('data_protection', None)
         self.is_restoring = kwargs.get('is_restoring', None)
-        self.snapshot_directory_visible = kwargs.get('snapshot_directory_visible', None)
+        self.snapshot_directory_visible = kwargs.get('snapshot_directory_visible', True)
         self.kerberos_enabled = kwargs.get('kerberos_enabled', False)
-        self.security_style = kwargs.get('security_style', None)
+        self.security_style = kwargs.get('security_style', "unix")
         self.smb_encryption = kwargs.get('smb_encryption', False)
         self.smb_continuously_available = kwargs.get('smb_continuously_available', False)
-        self.throughput_mibps = kwargs.get('throughput_mibps', None)
+        self.throughput_mibps = kwargs.get('throughput_mibps', 0)
+        self.encryption_key_source = kwargs.get('encryption_key_source', None)
 
 
 class VolumeBackupProperties(Model):